% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{extract_cpg_sites}
\alias{extract_cpg_sites}
\title{Extracts all CpGs from a genome}
\usage{
extract_cpg_sites(
  ref_genome = NULL,
  contigs = c(glue::glue("chr{1:22}"), "chrX", "chrY")
)
}
\arguments{
\item{ref_genome}{BSgenome object or name of the installed BSgenome package.
Example: BSgenome.Hsapiens.UCSC.hg19}
}
\value{
a list of data.table containing number of CpG's and contig lengths
}
\description{
Extracts all CpGs from a genome
}
