% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cpg_data.R
\name{read_cpg_data}
\alias{read_cpg_data}
\title{Read CpG data in MATT}
\usage{
read_cpg_data(
  fname,
  dataset_id = NULL,
  pipeline = "bedgraph",
  zero_based = TRUE,
  collapse_strands = TRUE,
  upper_cov_cutoff = list(method = "percentile", value = 0.999),
  reference = "hg19",
  align_to_reference = TRUE,
  compress = 100,
  use_cache = TRUE,
  update_cache = TRUE
)
}
\arguments{
\item{fname}{Path to data file.}

\item{dataset_id}{Unique identifier for CpG dataset in cache.}

\item{pipeline}{Pipeline for data processing. Supports "bedgraph", "gembs" or
custom format as a \code{list}.}

\item{zero_based}{Input file positions use 0-based indexing.}

\item{collapse_strands}{If \code{TRUE}, collapse forward and reverse strand data.}

\item{upper_cov_cutoff}{Filter for technical coverage errors.}

\item{reference}{Reference genome for CpG filling (alignment). Supports "hg19" and "hg38"}

\item{align_to_reference}{If \code{TRUE}, align or fill missing CpG sites based on reference.}

\item{compress}{Degree of compression for data cache (0-100).}

\item{use_cache}{If \code{TRUE}, save dataset to MATT cache using \code{dataset_id}}

\item{update_cache}{If \code{TRUE}, overwrite data in MATT cache using \code{dataset_id}}
}
\value{
A \code{data.table} with processed CpG data
}
\description{
Read CpG data in MATT
}
\examples{
infile <- system.file("extdata", "bismark_coverage_CpG.bedgraph", package="matt")
dt <- read_cpg_data(infile, align_to_reference=FALSE)
}
